Tab bar styling changes
The setting of WidgetBackground to VERY_DARK_GREY on the tab bar stack
will be inherited to all children. So, adding the WidgetBackground to
the Expand is redundant, because it already inherits that style.

The reason my Discord help wasn't working is that this WidetBackground
was being inherited and my example I was testing did not have that set.
So, to get the label to not draw a background, we set WidgetBackground
back to a clear color.