Use fluent-style APIs instead of arrays
The reason for the make_widgets is due to Rust not supporting multiple
types in an array. A tuple is perfect for this, but due to restrictions
in the Rust type system, I can't implement MakeWidgetList for tuples
*and* IntoIterator, because I get an error saying that IntoIterator
could be added someday to tuples.

Because I know this is a pain point, WidgetList is the type I created to
try to help solve these usability issues. The `widget.and(other_widget)`
syntax is the recommended way to avoid having to call make_widget when
creating lists of widgets.